backend/requirements.txt
# Web Framework
fastapi>=0.100.0
uvicorn>=0.22.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4

# Database
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# Utils
python-dotenv==1.0.0
pydantic>=2.0.0
pydantic-settings==2.1.0
alembic>=1.11.0
celery>=5.3.0
redis>=4.5.0
bcrypt==4.0.1 
pytesseract>=0.3.10
pdf2image>=1.16.3
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0

# AI/ML
openai>=1.0.0
requests>=2.31.0

# DWG/CAD
ezdxf>=1.0.0

# Visualization
matplotlib>=3.5.0

# Testing
pytest>=7.0.0
pytest-asyncio>=0.21.0 